Formatting

+++ src/main/java/com/minecraftabnormals/neapolitan/client/particle/ChimpanzeeNeedParticle.java
@@ -11,9 +11,9 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 public class ChimpanzeeNeedParticle extends SpriteTexturedParticle {
 	private ChimpanzeeNeedParticle(ClientWorld world, double x, double y, double z) {
 		super(world, x, y, z, 0.0D, 0.0D, 0.0D);
-		this.xd *= (double)0.01F;
-		this.yd *= (double)0.01F;
-		this.zd *= (double)0.01F;
+		this.xd *= 0.01F;
+		this.yd *= 0.01F;
+		this.zd *= 0.01F;
 		this.yd += 0.1D;
 		this.quadSize = 0.2F;
 		this.lifetime = 45;
@@ -25,7 +25,7 @@ public class ChimpanzeeNeedParticle extends SpriteTexturedParticle {
 	}
 
 	public float getQuadSize(float scaleFactor) {
-		return this.quadSize * MathHelper.clamp(((float)this.age + scaleFactor) / (float)this.lifetime * 32.0F, 0.0F, 1.0F);
+		return this.quadSize * MathHelper.clamp(((float) this.age + scaleFactor) / (float) this.lifetime * 32.0F, 0.0F, 1.0F);
 	}
 
 	public void tick() {
@@ -41,12 +41,12 @@ public class ChimpanzeeNeedParticle extends SpriteTexturedParticle {
 				this.zd *= 1.1D;
 			}
 
-			this.xd *= (double)0.86F;
-			this.yd *= (double)0.86F;
-			this.zd *= (double)0.86F;
+			this.xd *= 0.86F;
+			this.yd *= 0.86F;
+			this.zd *= 0.86F;
 			if (this.onGround) {
-				this.xd *= (double)0.7F;
-				this.zd *= (double)0.7F;
+				this.xd *= 0.7F;
+				this.zd *= 0.7F;
 			}
 		}
 	}
